pyproject.toml
[tool.poetry]
name = "destiny-transcript-db"
version = "0.1.0"
description = ""
authors = ["Ethan Rickert <ethanrickert0429@gmail.com>"]
readme = "README.md"

# packages = [{include = "destiny_transcript_db"}]
[tool.poetry.dependencies]
python = ">=3.8"
yt-dlp = "^2023.1.6"
python-dotenv = "^0.21.1"
internetarchive = "^3.3.0"
rich = "^13.3.1"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
